1. change the <Key> type from enum to int    in order to support some old compiler such as Clang-800.0.42.1 that doesn't enum key as a hashmap's key ( std::hash<A_CERTAIN_ENUM_TYPE> ) 2. add a pointer varible pointing to a default parser

+++ cpp/console/mathOperatation/core/parserException.cpp
@@ -9,7 +9,7 @@ ParserExpection::ParserExpection(E_ExceptionCode errorCode)
 	, m_formatDetail("")
 {
 	m_shortInfoMap.clear();
-	m_shortInfoMap.insert( make_pair(E_UNKNOWN_CHAR, string("Unknown Character") ) );
+	m_shortInfoMap.insert( make_pair(static_cast<int>(E_UNKNOWN_CHAR), string("Unknown Character") ) );
 }
 
 ParserExpection::ParserExpection(const  ParserExpection& r)
